requirements.txt
absl-py==2.3.1
accelerate==1.10.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
anywidget==0.9.18
asttokens==3.0.0
bbox_visualizer==0.2.2
black==25.1.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.3
decorator==5.2.1
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@18f69583391e5040043ca4f4bebd2c60f0ebfde0
einops==0.8.1
executing==2.2.0
fairscale==0.4.13
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.59.0
fsspec==2025.7.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
grpcio==1.74.0
h11==0.16.0
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.7
iopath==0.1.9
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_bbox_widget==0.6.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
omegaconf==2.3.0
open_clip_torch==3.1.0
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
peft==0.17.0
pi_heif==1.1.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
platformdirs==4.3.8
polygraphy==0.49.26
portalocker==3.2.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
psygnal==0.14.1
pure_eval==0.2.3
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pylabel==0.1.55
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
rf100vl==1.0.0
rfdetr==1.2.1
roboflow==1.2.6
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.2
supervision==0.26.1
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
timm==1.0.19
tldextract==5.3.0
tokenizers==0.21.4
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
yacs==0.1.8
